pyproject.toml
[project]
name = "djangoproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.0",
    "django>=6.0",
    "django-ckeditor>=6.7.3",
    "django-filter>=25.2",
    "django-unfold>=0.73.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular[sidecar]>=0.29.0",
    "markdown>=3.10",
    "pillow>=12.0.0",
    "redis>=7.1.0",
    "supabase>=2.27.1",
]
